requirements.txt
# Core Dependencies
pandas==2.0.3
numpy==1.24.3
scipy==1.11.1
scikit-learn==1.3.0

# Data Fetching
yfinance==0.2.28
pandas-datareader==0.10.0
requests==2.31.0

# Regime Detection
hmmlearn==0.3.0
statsmodels==0.14.0
ruptures==1.1.8  # Change-point detection

# Machine Learning & Optimization
stable-baselines3==2.0.0
optuna==3.3.0
torch==2.0.1  # Required by stable-baselines3

# Backtesting
vectorbt==0.25.0
ta==0.10.2  # Technical indicators

# Storage
duckdb==0.8.1
sqlalchemy==2.0.19

# Visualization & Reporting
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
quantstats==0.0.62
reportlab==4.0.4  # PDF generation
openpyxl==3.1.2  # Excel support

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Monitoring & Logging
prometheus-client==0.17.1
structlog==23.1.0

# Development & Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-asyncio==0.21.1
black==23.7.0
ruff==0.0.280
mypy==1.4.1
types-requests==2.31.0
types-PyYAML==6.0.12

# Documentation
sphinx==7.1.1
sphinx-rtd-theme==1.3.0

# Utilities
joblib==1.3.1
tqdm==4.65.0
click==8.1.6
python-dateutil==2.8.2